backend/pyproject.toml
[project]
name = "moveh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aptos-sdk>=0.7.0",
    "fastapi>=0.122.0",
    "langchain>=1.1.0",
    "langchain-core>=1.1.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    "modal>=1.2.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.5",
    "rich>=14.2.0",
    "tavily-python>=0.5.0",
    "uvicorn>=0.38.0",
]
